// frontend/src/app/eventoscaza/models/zona-marcada-impl.ts
import { Zonamarcada } from "./zonamarcada";

export class ZonaMarcadaImpl implements Zonamarcada {
  nombre: string;
  localidad: string;
  provincia: string;
  puntosImportantesHref: string;
  itinerarioCazaHref: string;
  coordLimitesZonaHref: string;
  observacionesZona: string;
  caza: string;

  constructor() {}


}